dsfc

+++ app/Models/Review.php
@@ -25,7 +25,7 @@ class Review extends Model
         'facilities_rating'=>'integer',
         'cleanliness_rating'=>'integer',
         'staff_rating'=>'integer',
-        'rating'=>'integer',
+        'rating'=>'double',
 
     ];
 }
